fix: prevent modules from producing same id hash
There was an obscure bug where the source code of 2 feed items was
totally identical which, when hashed, produced the same ids. Since we
can't have 2 different modules with the same id hash, we needed to add
another parameter to the hashing. I've simply taken the previous id
value (that generated by browserify) which is the path to the file
and prepended it to content to be hashed.